Add fifty.global.customElements
* Refactor testing documentation to be closer to testing implementation
* Make fifty.global more global in preparation for allowing tests to modify it
